requirements.txt
numpy==1.26.2
pandas==2.1.4
Pillow==10.3.0
scikit_learn==1.5.0
scikit-image==0.24.0
tensorflow==2.10.1
